+++ app/src/main/java/org/itstep/liannoi/fractioncalculator/application/fractions/models/Fraction.kt
@@ -1,6 +1,6 @@
 package org.itstep.liannoi.fractioncalculator.application.fractions.models
 
 data class Fraction constructor(
-    var numerator: Float = 0.0F,
-    var denominator: Float = 0.0F
+    var numerator: Int = 0,
+    var denominator: Int = 0
 )
